boards: allow for GITHUB_LABEL variable
If the GITHUB_LABEL variable is set, then it is a semicolon-separated
list of labels to consider for Azure Boards work items.  If it is set
then the GitHub issue must have a label within the GITHUB_LABEL variable
to be created in Azure Boards.  If it remains unset then all GitHub
issues will have an Azure Boards work item created.