pipelines: stylistic changes in env var tests
Minor stylistic changes in the environment variable tests: first, in
error messages, do not include the leading `$`.  It's uncommon to
include in error messages and is shell-dependent (PowerShell, eg,
uses a different syntax).  Removal standardizes the messages with the
other scripts in this repository.

Also, do not keep a trailing semicolon on the `if` conditional, which
delimits commands.  The semicolon should either be removed if the `then`
is on a newline, or the `then` should be on the same line as the `if`
(to make the semicolon functional).  I've moved the `then` up to the
line with the `if` to standardize with the other scripts in this
repository.