boards: use AZURE_DEVOPS_ variables for configuration
Instead of using `AZURE_BOARDS_` variables for the common configuration
to Azure DevOps, prefix the names with `AZURE_DEVOPS_`.  This allows
users to re-use configuration between multiple instances of GitHub
Actions that target different Azure DevOps services.

This is especially useful for personal access tokens, which are secret
variables and thus not copy/pasteable between GitHub Actions
configurations, but *do* default to be shared.  If the shared secret
variable is `AZURE_DEVOPS_TOKEN` then both boards *and* pipelines
actions can use it without additional configuration or a second token.